Add support for manual id.

+++ scripts/collector/postgres/collect-data.sh
@@ -108,11 +108,11 @@ function checkVersionPg {
     fi
 
     # SELECT 'DMAFILETAG~' , version();
-    dbversion=$(${SQLCMD}  --user=$user --password -h $host -w -p $port -t << EOF
+    dbversion=$(${SQLCMD}  --user=$user --password -h $host -w -p $port -t --no-align << EOF
 SELECT current_setting('server_version');
 EOF
 )
-echo 'DMAFILETAG~'${dbversion}'|'${dbversion}'_'${OpVersion}'_'${host}'-'${port}'_'${db}'_'${db}
+echo 'DMAFILETAG~'${dbversion}'|'${dbversion}'_'${OpVersion}'_'${host}'-'${port}'_'${db}'_'${db}'_'$(date +%y%m%d%H%M%S)
 }
 
 function checkVersionMysql {
@@ -132,7 +132,6 @@ function checkVersionMysql {
       exit 1
     fi
 
-# SELECT 'DMAFILETAG~' , version();
 dbversion=$(${SQLCMD}  --user=$user --password=$pass -h $host -P $port -s $db << EOF
 SELECT version();
 EOF
@@ -185,10 +184,12 @@ EOF
 }
 
 
+
 function executeOPMysql {
 connectString="$1"
 OpVersion=$2
 V_FILE_TAG=$3
+V_MANUAL_ID="${4}"
 user=$(echo ${connectString} | cut -d '/' -f 1)
 pass=$(echo ${connectString} | cut -d '/' -f 2 | cut -d '@' -f 1)
 host=$(echo ${connectString} | cut -d '/' -f 4 | cut -d ':' -f 1)
@@ -203,13 +204,22 @@ fi
 
 export SKIPSCHEMA=$(grep -v \# sql/source/skipschema.csv)
 
+echo Getting DMA_SOURCE_ID with  ${SQLCMD} --user=$user --password=$pass -h $host -P $port --force --table --silent $db 
+export DMA_SOURCE_ID=$(${SQLCMD} --user=$user --password=$pass -h $host -P $port --force --silent --skip-column-names $db < sql/source/init.sql | tr -d '
+')
+echo DMA_SOURCE_ID = $DMA_SOURCE_ID
+
 for s in sql/source/*sql
 do
   fname=$(echo $s | cut -d '/' -f 3)
-  ${SED} "s/V_TAG/${V_TAG}/g;s/SKIPSCHEMA/${SKIPSCHEMA}/g;s/SQLOUTPUT_DIR/'${SQLOUTPUT_DIR}'/g" ${s} > sql/${V_FILE_TAG}_${fname}
+  ${SED} "s/V_TAG/${V_TAG}/g;s/SKIPSCHEMA/${SKIPSCHEMA}/g;s/SQLOUTPUT_DIR/'${SQLOUTPUT_DIR}'/g;s/_DMASOURCEID_/${DMA_SOURCE_ID}/g;s/_DMAMANUALID_/${V_MANUAL_ID}/g" ${s} > sql/${V_FILE_TAG}_${fname}
 done
 
+if [ -f sql/${V_TAG}_mysqlcollector.sql ]; 
+then
 rm sql/${V_TAG}_mysqlcollector.sql
+fi
+
 for f in sql/${V_FILE_TAG}_*sql
 do
   echo source ${f} >> sql/${V_FILE_TAG}_mysqlcollector.sql
@@ -217,6 +227,11 @@ done
 
 ${SQLCMD} --user=$user --password=$pass -h $host -P $port --force --table $db < sql/${V_FILE_TAG}_mysqlcollector.sql
 
+for x in $(grep -L DMA_SOURCE_ID output/*${V_FILE_TAG}.csv )
+do
+  sed 's/
+//g' ${x} | awk -v SRCID="${DMA_SOURCE_ID}" -v MANID="${V_MANUAL_ID}" -v Q="'" 'BEGIN { FS="|"; OFS="|"; }; {if (NR == 2) { $NF = $NF "DMA_SOURCE_ID" "|" "MANUAL_ID" } else { $NF = $NF Q SRCID Q "|" Q MANID Q "|" } print }' > ${x}.tmp && mv ${x}.tmp ${x}
+done
 }
 
 
@@ -224,6 +239,7 @@ function executeOPPg {
 connectString="$1"
 OpVersion=$2
 V_FILE_TAG=$3
+V_MANUAL_ID="${4}"
 user=$(echo ${connectString} | cut -d '/' -f 1)
 pass=$(echo ${connectString} | cut -d '/' -f 2 | cut -d '@' -f 1)
 host=$(echo ${connectString} | cut -d '/' -f 4 | cut -d ':' -f 1)
@@ -237,14 +253,19 @@ if ! [ -x "$(command -v ${SQLCMD})" ]; then
 fi
 
 
-DMA_SOURCE_ID=$(${SQLCMD}  --user=$user --password -h $host -w -p $port -t <<EOF
+DMA_SOURCE_ID=$(${SQLCMD}  --user=$user --password -h $host -w -p $port -t --no-align <<EOF
 SELECT system_identifier FROM pg_control_system();
 EOF
 )
+if [[ "${V_MANUAL_ID}" == "" ]]
+then
+	V_MANUAL_ID=""
+fi
 
-${SQLCMD}  --user=$user --password -h $host -w -p $port -t <<EOF
+${SQLCMD}  --user=$user --password -h $host -w -p $port -t --no-align <<EOF
 \set VTAG ${V_FILE_TAG}
-\set DMA_SOURCE_ID ${DMA_SOURCE_ID}
+\set DMA_SOURCE_ID '\'${DMA_SOURCE_ID}\''
+\set DMA_MANUAL_ID '\'${V_MANUAL_ID}\''
 \i sql/op_collect.sql
 EOF
 }
@@ -318,6 +339,7 @@ ZIPFILE=opdb_${DBTYPE}_${DIAGPACKACCESS}__${V_FILE_TAG}${V_ERR_TAG}.zip
 
 locale > ${OUTPUT_DIR}/opdb__${V_FILE_TAG}_locale.txt
 
+echo "MANUAL_ID : " ${MANUAL_ID} >> ${OUTPUT_DIR}/opdb__defines__${V_FILE_TAG}.csv
 echo "ZIPFILE: " $ZIPFILE >> ${OUTPUT_DIR}/opdb__defines__${V_FILE_TAG}.csv
 
 cd ${OUTPUT_DIR}
@@ -345,6 +367,10 @@ fi
 if [ -f $OUTFILE ]
 then
   rm opdb*${V_FILE_TAG}.csv opdb*${V_FILE_TAG}*.log opdb*${V_FILE_TAG}*.txt
+  if [[ -f ../sql/${V_FILE_TAG}*.sql ]]
+  then
+    rm ../sql/${V_FILE_TAG}*.sql
+  fi
 fi
 
 cd ${CURRENT_WORKING_DIR}
@@ -386,7 +412,7 @@ fi
 ### Validate input
 #############################################################################
 
-if [[  $# -ne 3  || (  "$2" != "UseDiagnostics" && "$2" != "NoDiagnostics" ) ]]
+if [[  $# -lt 3  || $# -gt 4 || (  "$2" != "UseDiagnostics" && "$2" != "NoDiagnostics" ) ]]
  then
   echo
   echo "You must indicate whether or not to use the Diagnostics Pack views."
@@ -402,9 +428,15 @@ fi
 # MAIN
 #############################################################################
 
-DBTYPE="$3"
 
 connectString="$1"
+DIAGPACKACCESS="$2"
+DBTYPE="$3"
+MANUALID=""
+if [[ $# -eq 4 ]]; 
+then
+  MANUALID=$(echo "$4" | iconv -t ascii//TRANSLIT | sed -E -e 's/[^[:alnum:]]+/-/g' -e 's/^-+|-+$//g' | tr '[:upper:]' '[:lower:]')
+fi
 
 checkPlatform $DBTYPE
 
@@ -418,7 +450,6 @@ if [ "$DBTYPE" == "oracle" ] ; then
     fi
   else if [ "$DBTYPE" == "mysql" ] ; then
     sqlcmd_result=$(checkVersionMysql "${connectString}" "${OpVersion}" | $GREP DMAFILETAG | tr -d ' ' | cut -d '~' -f 2 | tr -d '\r' )
-    echo Result=${sqlcmd_result}
     if [[ "${sqlcmd_result}" = "" ]];
       then
       echo "Unable to connect to the target MySQL database using ${connectString}.  Please verify the connection information and target database status."
@@ -426,7 +457,6 @@ if [ "$DBTYPE" == "oracle" ] ; then
     fi
     else if [ "$DBTYPE" == "postgres" ] ; then
       sqlcmd_result=$(checkVersionPg "${connectString}" "${OpVersion}" | $GREP DMAFILETAG | tr -d ' ' | cut -d '~' -f 2 | tr -d '\r' )
-      echo Result=${sqlcmd_result}
       if [[ "${sqlcmd_result}" = "" ]];
         then
         echo "Unable to connect to the target Postgres database using ${connectString}.  Please verify the connection information and target database status."
@@ -438,7 +468,6 @@ fi
 
 retval=$?
 
-DIAGPACKACCESS="$2"
 
 extractorVersion="$(getVersion)"
 
@@ -463,16 +492,15 @@ if [ $retval -eq 0 ]; then
        echo "Oracle 10 support is experimental."
     fi
     V_TAG="$(echo ${sqlcmd_result} | cut -d '|' -f2).csv"; export V_TAG
-    echo V_TAG = ${V_TAG}
 
     if [ "$3" == "oracle" ] ; then
       executeOPOracle "${connectString}" ${OpVersion} ${DIAGPACKACCESS}
       retval=$?
     else if [ "$3" == "mysql" ]; then
-      executeOPMysql "${connectString}" ${OpVersion} $(echo ${V_TAG} | ${SED} 's/.csv//g')
+      executeOPMysql "${connectString}" ${OpVersion} $(echo ${V_TAG} | ${SED} 's/.csv//g') "${MANUALID}"
       retval=$?
     else if [ "$3" == "postgres" ]; then
-      executeOPPg "${connectString}" ${OpVersion} $(echo ${V_TAG} | ${SED} 's/.csv//g')
+      executeOPPg "${connectString}" ${OpVersion} $(echo ${V_TAG} | ${SED} 's/.csv//g') "${MANUALID}"
       retval=$?
       fi
     fi
